+++ src/main/java/projet/projetstage02/controller/RootController.java
@@ -8,31 +8,18 @@ import org.apache.logging.log4j.Logger;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import projet.projetstage02.dto.SignatureInDTO;
-import projet.projetstage02.dto.applications.ApplicationAcceptationDTO;
-import projet.projetstage02.dto.applications.ApplicationDTO;
-import projet.projetstage02.dto.applications.ApplicationListDTO;
-import projet.projetstage02.dto.applications.OfferApplicationDTO;
 import projet.projetstage02.dto.auth.LoginDTO;
 import projet.projetstage02.dto.auth.TokenDTO;
 import projet.projetstage02.dto.contracts.ContractsDTO;
 import projet.projetstage02.dto.contracts.StageContractInDTO;
 import projet.projetstage02.dto.contracts.StageContractOutDTO;
 import projet.projetstage02.dto.cv.CvRefusalDTO;
-import projet.projetstage02.dto.cv.CvStatusDTO;
 import projet.projetstage02.dto.evaluations.MillieuStage.MillieuStageEvaluationInDTO;
 import projet.projetstage02.dto.evaluations.MillieuStage.MillieuStageEvaluationInfoDTO;
-import projet.projetstage02.dto.interview.CreateInterviewDTO;
-import projet.projetstage02.dto.interview.InterviewOutDTO;
-import projet.projetstage02.dto.interview.InterviewSelectInDTO;
-import projet.projetstage02.dto.offres.OfferAcceptedStudentsDTO;
-import projet.projetstage02.dto.offres.OffreInDTO;
 import projet.projetstage02.dto.offres.OffreOutDTO;
-import projet.projetstage02.dto.pdf.PdfDTO;
 import projet.projetstage02.dto.pdf.PdfOutDTO;
 import projet.projetstage02.dto.users.CompanyDTO;
 import projet.projetstage02.dto.users.GestionnaireDTO;
-import projet.projetstage02.dto.users.Students.StudentInDTO;
 import projet.projetstage02.dto.users.Students.StudentOutDTO;
 import projet.projetstage02.exception.*;
 import projet.projetstage02.model.Token;
@@ -67,31 +54,7 @@ public class RootController {
 
 
 
-    @PostMapping("/createCompany")
-    public ResponseEntity<Map<String, String>> createCompany(@Valid @RequestBody CompanyDTO companyDTO) {
-        try {
-            logger.log(INFO, "Post /createCompany entered with body : " + companyDTO.toString());
-            if (companyService.isCompanyInvalid(companyDTO.getEmail())) {
-                logger.log(INFO, "PostMapping: /createCompany sent 409 response");
-                return ResponseEntity.status(CONFLICT)
-                        .body(getError("Cette adresse email est déjà utilisée."));
-            }
-            companyDTO.setInscriptionTimestamp(currentTimestamp());
-            long id = companyService.saveCompany(companyDTO);
-            companyDTO.setId(id);
-            if (!EmailUtil.sendConfirmationMail(companyDTO.toModel())) {
-                logger.log(INFO, "PostMapping: /createCompany sent 500 response");
-                return ResponseEntity.status(INTERNAL_SERVER_ERROR).body(
-                        getError("Une erreur avec le service d'email est survenue"));
-            }
-            logger.log(INFO, "PostMapping: /createCompany sent 201 response");
-            return ResponseEntity.status(CREATED).build();
-        } catch (NonExistentEntityException e) {
-            logger.log(INFO, "PostMapping: /createCompany sent 500 response");
-            return ResponseEntity.status(INTERNAL_SERVER_ERROR).body(
-                    getError("Une erreur est survenue"));
-        }
-    }
+
 
     @SneakyThrows
     @PostMapping("/createGestionnaire")
@@ -124,21 +87,7 @@ public class RootController {
         }
     }
 
-    @PostMapping("/createOffre")
-    public ResponseEntity<Map<String, String>> createOffre(@Valid @RequestBody OffreInDTO offreInDTO) {
-        try {
-            logger.log(INFO, "Post /createOffre entered with body : " + offreInDTO);
-            Token token = authService.getToken(offreInDTO.getToken(), COMPANY);
-            offreInDTO.setSession("Hiver " + getNextYear());
-            offreInDTO.setCompanyId(token.getUserId());
-            companyService.createOffre(offreInDTO);
-            logger.log(INFO, "PostMapping: /createOffre sent 201 response");
-            return ResponseEntity.status(CREATED).build();
-        } catch (InvalidTokenException ex) {
-            logger.log(INFO, "PostMapping: /createOffre sent 403 response");
-            return ResponseEntity.status(FORBIDDEN).build();
-        }
-    }
+
 
     private Map<String, String> getError(String error) {
         HashMap<String, String> hashMap = new HashMap<>();
@@ -148,25 +97,7 @@ public class RootController {
 
 
 
-    @PutMapping("/confirmEmail/company/{id}")
-    public ResponseEntity<Map<String, String>> confirmCompanyMail(@PathVariable String id) {
-        logger.log(INFO, "Put /confirmEmail/company/{id} entered with id : " + id);
-        try {
-            CompanyDTO companyDTO = companyService.getCompanyById(Long.parseLong(id));
-            if (currentTimestamp() - companyDTO.getInscriptionTimestamp() > MILLI_SECOND_DAY) {
-                logger.log(INFO, "PutMapping: /confirmEmail/company sent 400 response");
-                return ResponseEntity.status(BAD_REQUEST)
-                        .body(getError("La période de confirmation est expirée"));
-            }
-            companyDTO.setEmailConfirmed(true);
-            companyService.saveCompany(companyDTO);
-            logger.log(INFO, "PutMapping: /confirmEmail/company sent 201 response");
-            return ResponseEntity.status(CREATED).build();
-        } catch (NonExistentEntityException e) {
-            logger.log(INFO, "PutMapping: /confirmEmail/company sent 404 response");
-            return ResponseEntity.notFound().build();
-        }
-    }
+
 
     @PutMapping("/confirmEmail/gestionaire/{id}")
     public ResponseEntity<Map<String, String>> confirmGestionnaireMail(@PathVariable String id) {
@@ -228,25 +159,7 @@ public class RootController {
 
 
 
-    @PutMapping("/company")
-    public ResponseEntity<CompanyDTO> getCompany(@Valid @RequestBody TokenDTO tokenId) {
-        logger.log(INFO, "Put /company entered with body : " + tokenId.toString());
-        try {
-            Token token = authService.getToken(tokenId.getToken(), COMPANY);
 
-            CompanyDTO dto = companyService.getCompanyById(token.getUserId());
-            dto.setPassword("");
-            logger.log(INFO, !dto.isEmailConfirmed() ? "Put /company entered sent 400 response"
-                    : "Put /company entered sent 200 response");
-            return !dto.isEmailConfirmed() ? ResponseEntity.notFound().build() : ResponseEntity.ok(dto);
-        } catch (NonExistentEntityException e) {
-            logger.log(INFO, "Put /company entered sent 404 response");
-            return ResponseEntity.notFound().build();
-        } catch (InvalidTokenException ex) {
-            logger.log(INFO, "PutMapping: /company sent 403 response");
-            return ResponseEntity.status(FORBIDDEN).build();
-        }
-    }
 
     @PutMapping("/gestionnaire")
     public ResponseEntity<GestionnaireDTO> getGestionnaire(@Valid @RequestBody TokenDTO tokenId) {
@@ -486,20 +399,7 @@ public class RootController {
         }
     }
 
-    @PutMapping("/company/studentCv/{studentId}")
-    private ResponseEntity<PdfOutDTO> getCvAsCompany(@PathVariable long studentId, @RequestBody TokenDTO tokenId) {
-        try {
-            authService.getToken(tokenId.getToken(), COMPANY);
-            PdfOutDTO cv = companyService.getStudentCv(studentId);
-            logger.log(INFO, "PutMapping: /studentCv sent 200 response");
-            return ResponseEntity.ok(cv);
-        } catch (InvalidTokenException ex) {
-            logger.log(INFO, "PutMapping: /studentCv sent 403 response");
-            return ResponseEntity.status(FORBIDDEN).build();
-        } catch (NonExistentEntityException ex) {
-            return ResponseEntity.notFound().build();
-        }
-    }
+
 
     @PutMapping("/validateCv/{studentId}")
     public ResponseEntity<StudentOutDTO> validateStudentCv(@PathVariable String studentId, @RequestBody TokenDTO tokenId) {
@@ -536,51 +436,9 @@ public class RootController {
         }
     }
 
-    @PutMapping("/studentAcceptation/{offerId}_{studentId}")
-    public ResponseEntity<ApplicationAcceptationDTO> saveStudentAcceptation
-            (@PathVariable String offerId, @PathVariable String studentId, @RequestBody TokenDTO tokenId) {
-        logger.log(INFO, "Put /studentAcceptation/{offerId}_{studentId} entered with offerId: " + offerId
-                + " and studentId: " + studentId);
 
-        try {
-            authService.getToken(tokenId.getToken(), COMPANY);
-            ApplicationAcceptationDTO dto = companyService
-                    .saveStudentApplicationAccepted(Long.parseLong(offerId), Long.parseLong(studentId));
-            logger.log(INFO, "Put /studentAcceptation/{offerId}_{studentId} sent 200 response");
-            return ResponseEntity.ok(dto);
-        } catch (InvalidTokenException e) {
-            logger.log(INFO, "Put /studentAcceptation/{offerId}_{studentId} sent 403 response");
-            return ResponseEntity.status(FORBIDDEN).build();
-        } catch (AlreadyExistingAcceptationException e) {
-            logger.log(INFO, "Put /studentAcceptation/{offerId}_{studentId} sent 409 response");
-            return ResponseEntity.status(CONFLICT).build();
-        } catch (NonExistentEntityException e) {
-            logger.log(INFO, "Put /studentAcceptation/{offerId}_{studentId} sent 404 response");
-            return ResponseEntity.notFound().build();
-        } catch (NonExistentOfferExeption e) {
-            logger.log(INFO, "Put /studentAcceptation/{offerId}_{studentId} sent 404 response : ");
-            return ResponseEntity.notFound().build();
-        }
-    }
 
-    @PutMapping("/getAcceptedStudentsForOffer/{offerId}")
-    public ResponseEntity<OfferAcceptedStudentsDTO> getAcceptedStudentsForOffer
-            (@PathVariable String offerId, @RequestBody TokenDTO tokenId) {
-        logger.log(INFO, "Put /getAcceptedStudentsForOffer/{offerId} entered with offerId: " + offerId);
 
-        try {
-            authService.getToken(tokenId.getToken(), COMPANY);
-            OfferAcceptedStudentsDTO dto = companyService.getAcceptedStudentsForOffer(Long.parseLong(offerId));
-            logger.log(INFO, "Put /getAcceptedStudentsForOffer/{offerId} sent 200 response");
-            return ResponseEntity.ok(dto);
-        } catch (InvalidTokenException e) {
-            logger.log(INFO, "Put /getAcceptedStudentsForOffer/{offerId} sent 403 response");
-            return ResponseEntity.status(FORBIDDEN).build();
-        } catch (NonExistentOfferExeption e) {
-            logger.log(INFO, "Put /getAcceptedStudentsForOffer/{offerId} sent 404 response");
-            return ResponseEntity.notFound().build();
-        }
-    }
 
     //todo replace .log with .info
     @PostMapping("/createStageContract")
@@ -604,44 +462,7 @@ public class RootController {
         }
     }
 
-    @PutMapping("/companyContracts/{companyId}_{session}")
-    public ResponseEntity<ContractsDTO> getCompanyContracts
-            (@PathVariable String companyId, @PathVariable String session, @RequestBody TokenDTO tokenId) {
-        logger.log(INFO, "Put /companyContracts/{companyId} entered with companyId: " + companyId
-                + " with the session: " + session);
 
-        try {
-            authService.getToken(tokenId.getToken(), COMPANY);
-            ContractsDTO contracts = companyService.getContracts(Long.parseLong(companyId), session);
-            logger.log(INFO, "Put /companyContracts/{companyId} sent 200 response");
-            return ResponseEntity.ok(contracts);
-        } catch (NonExistentEntityException e) {
-            logger.log(INFO, "Put /companyContracts/{companyId}_{session} sent 404 response");
-            return ResponseEntity.notFound().build();
-        } catch (InvalidTokenException e) {
-            logger.log(INFO, "Put /companyContracts/{companyId}_{session} sent 403 response");
-            return ResponseEntity.status(FORBIDDEN).build();
-        }
-    }
-
-    @PutMapping("/companySignatureContract")
-    public ResponseEntity<StageContractOutDTO> companyContractSignature
-            (@Valid @RequestBody SignatureInDTO signatureInDTO) {
-        logger.log(INFO, "Put /companySignatureContract entered with SignatureInDTO: " + signatureInDTO);
-
-        try {
-            authService.getToken(signatureInDTO.getToken(), COMPANY);
-            StageContractOutDTO dto = companyService.addSignatureToContract(signatureInDTO);
-            logger.log(INFO, "Put /companySignatureContract sent request 200 : " + dto);
-            return ResponseEntity.ok(dto);
-        } catch (NonExistentEntityException e) {
-            logger.log(INFO, "Put /companySignatureContract sent request 404");
-            return ResponseEntity.notFound().build();
-        } catch (InvalidTokenException | InvalidOwnershipException e) {
-            logger.log(INFO, "Put /companySignatureContract sent request 403");
-            return ResponseEntity.status(FORBIDDEN).build();
-        }
-    }
 
 
 
@@ -662,39 +483,7 @@ public class RootController {
         }
     }
 
-    @PutMapping("/offer/{id}/applications")
-    public ResponseEntity<OfferApplicationDTO> getStudentForOffer
-            (@PathVariable long id, @RequestBody TokenDTO tokenId) {
-        logger.log(INFO, "/offer/{id}/applications entered with id: " + id);
-        try {
-            authService.getToken(tokenId.getToken(), COMPANY);
-            OfferApplicationDTO dto = companyService.getStudentsForOffer(id);
-            logger.log(INFO, "/offer/{id}/applications sent 200 response");
-            return ResponseEntity.ok(dto);
-        } catch (InvalidTokenException e) {
-            logger.log(INFO, "/offer/{id}/applications sent 403 response");
-            return ResponseEntity.status(FORBIDDEN).build();
-        } catch (NonExistentOfferExeption | NonExistentEntityException e) {
-            logger.log(INFO, "/offer/{id}/applications sent 404 response");
-            return ResponseEntity.notFound().build();
-        }
-    }
 
-    @PutMapping("/company/validatedOffers/{id}")
-    //todo replace tokens with @RequestHeader("Authorization") String token
-    public ResponseEntity<List<OffreOutDTO>> getValidatedOffers
-            (@PathVariable long id, @RequestBody TokenDTO tokenId) {
-        try {
-            logger.log(INFO, "put /validatedOffers entered with id : " + id);
-            authService.getToken(tokenId.getToken(), COMPANY);
-            List<OffreOutDTO> validatedOffers = companyService.getValidatedOffers(id);
-            logger.log(INFO, "PutMapping: /validatedOffers sent 200 response");
-            return ResponseEntity.ok(validatedOffers);
-        } catch (InvalidTokenException e) {
-            logger.log(INFO, "PutMapping: /validatedOffers sent 403 response");
-            return ResponseEntity.status(FORBIDDEN).build();
-        }
-    }
 
     @PutMapping("/evaluateStage/{contractId}/getInfo")
     public ResponseEntity<MillieuStageEvaluationInfoDTO> getEvaluateStageInfo
@@ -745,45 +534,7 @@ public class RootController {
         }
     }
 
-    @PostMapping("/createInterview")
-    public ResponseEntity<InterviewOutDTO> createInterview(@RequestBody @Valid CreateInterviewDTO interviewDTO){
-        logger.log(INFO, "Post /createInterview");
 
-        try {
-            authService.getToken(interviewDTO.getToken(), COMPANY);
-            InterviewOutDTO dto = companyService.createInterview(interviewDTO);
-            logger.log(INFO, "Post /createInterview return 201 request");
-            return ResponseEntity.status(CREATED).body(dto);
-        } catch (NonExistentEntityException e) {
-            logger.log(INFO, "Post /createInterview return 404 request");
-            return ResponseEntity.status(NOT_FOUND).build();
-        } catch (InvalidDateFormatException  e) {
-            logger.log(INFO, "Post /createInterview return 400 request");
-            return ResponseEntity.status(BAD_REQUEST).build();
-        } catch (InvalidTokenException | InvalidOwnershipException e) {
-            logger.log(INFO, "Post /createInterview return 403 request");
-            return ResponseEntity.status(FORBIDDEN).build();
-        }
-    }
-
-    @PutMapping("/getCompanyInterviews/{companyId}")
-    public ResponseEntity<List<InterviewOutDTO>> getCompanyInterviews
-            (@PathVariable String companyId, @RequestBody TokenDTO tokenId){
-        logger.log(INFO, "Put /getCompanyInterviews/{companyId}");
-
-        try {
-            authService.getToken(tokenId.getToken(), COMPANY);
-            List<InterviewOutDTO> interviews = companyService.getInterviews(Long.parseLong(companyId));
-            logger.log(INFO, "Put /getCompanyInterviews/{companyId} return 200");
-            return ResponseEntity.ok(interviews);
-        } catch (NonExistentEntityException e) {
-            logger.log(INFO, "Put /getCompanyInterviews/{companyId} return 404");
-            return ResponseEntity.status(NOT_FOUND).build();
-        } catch (InvalidTokenException e) {
-            logger.log(INFO, "Put /getCompanyInterviews/{companyId} return 403");
-            return ResponseEntity.status(FORBIDDEN).build();
-        }
-    }
 
 
 }
